LayoutTests/fast/text/system-font-fallback-emoji.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div style="font-size: 100px;">
<div><span id="systemFontTarget" style="font-family: 'system-ui', 'AppleColorEmoji';">😊</span></div>
<div><span id="appleColorEmojiTarget" style="font-family: 'AppleColorEmoji';">😊</span></div>
</div>
<script>
description("This test makes sure that the system font draws emoji correctly.");
var systemFontTarget = document.getElementById("systemFontTarget");
var appleColorEmojiTarget = document.getElementById("appleColorEmojiTarget");
shouldNotBe("systemFontTarget.offsetWidth", "100");
shouldBe("appleColorEmojiTarget.offsetWidth", "100");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
