Safari-609.1.11.1/LayoutTests/fast/text/system-font-fallback-emoji-expected.txt

This test makes sure that the system font draws emoji correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS systemFontTarget.offsetWidth is not 100
PASS appleColorEmojiTarget.offsetWidth is 100
PASS successfullyParsed is true

TEST COMPLETE
😊
😊
